fix: make und work in g2p studio
The reason und did not work (see issue #165) is that class Transducer
relies on having type == "unidecode" to trigger the und unidecode
processing.

I don't really like this solution that much, because it makes this extra
parameter visible to all mappings, just to support one mapping. I'm open
to suggestions for better solutions. Maybe that field can be hidden?